pyproject.toml
[tool.poetry]
name = "githubCodeownerQuery"
version = "0.1.0"
description = ""
authors = ["Sedthakit Prsanphanich <sedthakit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^1.59.1"


[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
